Sp reservation list screen (#5)
* add topbar and avatar

* feat: Update car wash package prices and services

The car wash package prices and services have been updated to reflect the changes made in the JSON files. This ensures that the prices and services displayed on the HomeScreen are accurate and up to date.

* Update selectReservations selector to access reservations state in reservationSlicens

* feat: Update reservation list screen

Refactor ReservationListScreen component to include a "Go Back Home" button for easy navigation. Also update the header text to "My Reservations" for clarity.